Fix type signature
iso_to_timestamp accepts all types, but prefers strings, for
documentational purposes I put this as Optional[str] instead of Any or
object.

call_rems_api returns a list of permissions (=> add List). Exceptions
may be raised, but this should not be shown by the typse (=> remove
Optional). The same goes for the input to create_ga4gh_visa_v1.